Fixed: everything inside blocks and no logging

+++ opsworks_rubber/recipes/deploy.rb
@@ -11,26 +11,20 @@ opsworks_deploy_dir do
   path deploy[:deploy_to]
 end
 
-Chef::Log.info('Deploying Rubber Application')
 opsworks_deploy do
   deploy_data deploy
   app application
 end
 
-if File.exists?("#{deploy[:current_path]}/Gemfile")
-  Chef::Log.info("Gemfile detected. Running bundle install.")
-  cmd = "sudo su - #{deploy[:user]} -c 'cd #{deploy[:current_path]} && /usr/local/bin/bundle install --path #{deploy[:home]}/.bundler/#{application} '"
-  Chef::Log.info(cmd)
-  bash 'bundle_install' do
-    user deploy[:user]
-    cwd deploy[:current_path]
-    code <<-EOH
-      /usr/local/bin/bundle install --path #{deploy[:home]}/.bundler/#{application} --without=#{deploy[:ignore_bundler_groups].join(' ')}
-    EOH
-  end
+bash 'bundle_install' do
+  user deploy[:user]
+  cwd deploy[:current_path]
+  code <<-EOH
+    /usr/local/bin/bundle install --path #{deploy[:home]}/.bundler/#{application} --without=#{deploy[:ignore_bundler_groups].join(' ')}
+  EOH
+  only_if { File.exists?("#{deploy[:current_path]}/Gemfile") }
 end
 
-Chef::Log.info('Restarting Rubber Application')
 execute 'restart Rubber app' do
   command node[:rubber][:commands][:restart]
 end
